+++ app/auth/Nav.tsx
@@ -1,11 +1,23 @@
 import Link from "next/link";
+import Login from "./Login";
+import Logged from "./Logged";
+import { getServerSession } from "next-auth/next";
+import { authOptions } from "../../pages/api/auth/[...nextauth]";
 
 export default async function Nav() {
+  const session = await getServerSession(authOptions);
+  console.log(session);
+
   return (
-    <nav>
+    <nav className="flex justify-between items-center py-8">
       <Link href={"/"}>
-        <h1>Send this</h1>
+        <h1 className="font-bold text-xl">GroupUp</h1>
       </Link>
+      {/* client componenet */}
+      <ul className="flex items-center gap-6">
+        {!session?.user && <Login />}
+        {session?.user && <Logged image={session.user?.image || ""} />}
+      </ul>
     </nav>
   );
 }