+++ app/components/CreatePost.tsx
@@ -1,0 +1,67 @@
+"use client";
+
+import { useState } from "react";
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+import axios, { AxiosError } from "axios";
+import toast from "react-hot-toast";
+
+export default function CreatePost() {
+  const [title, setTitle] = useState("");
+  const [isDisabled, setIsDisabled] = useState(false);
+  let toastPostId: string;
+
+  // create post
+  const { mutate } = useMutation(
+    async (title: string) =>
+      await axios.post("/api/posts/createPost", { title }),
+    {
+      onError: (error) => {
+        if (error instanceof AxiosError) {
+          toast.error(error?.response?.data.message, { id: toastPostId });
+        }
+        setIsDisabled(false);
+      },
+      onSuccess: (data) => {
+        toast.success("Post created!", { id: toastPostId });
+
+        setTitle("");
+        setIsDisabled(false);
+      },
+    }
+  );
+
+  const submitPost = async (e: React.FormEvent) => {
+    e.preventDefault();
+    toastPostId = toast.loading("Creating post...", { id: toastPostId });
+    setIsDisabled(true);
+    mutate(title);
+  };
+
+  return (
+    <form onSubmit={submitPost} className="bg-white my-8 p-8 rounded-md">
+      <div className="flex flex-col my-4">
+        <textarea
+          className="p-4 text-lg rounded-md my-t bg-gray-200"
+          name="title"
+          placeholder="Write some stuff"
+          value={title}
+          onChange={(e) => setTitle(e.target.value)}
+        ></textarea>
+      </div>
+      <div className="flex items-center justify-between gap-2">
+        <p
+          className={`font-bold text-sm ${
+            title.length > 300 ? "text-red-700" : "text-gray-700"
+          }`}
+        >{`${title.length}/300`}</p>
+        <button
+          className="text-sm bg-teal-600 text-white p-2 rounded-xl disabled:opacity-25"
+          disabled={isDisabled}
+          type="submit"
+        >
+          Create a post
+        </button>
+      </div>
+    </form>
+  );
+}