+++ app/page.tsx
@@ -1,7 +1,38 @@
+"use client";
+
+import axios from "axios";
+import CreatePost from "./components/CreatePost";
+import { useQuery } from "@tanstack/react-query";
+import Posts from "./components/Posts";
+import { PostType } from "./types/Posts";
+
+// get all posts
+const getAllPosts = async () => {
+  const response = await axios.get("/api/posts/getPosts");
+  return response.data;
+};
+
 export default function Home() {
+  const { data, error, isLoading } = useQuery<PostType[]>({
+    queryFn: getAllPosts,
+    queryKey: ["posts"],
+  });
+  if (error) return error;
+  if (isLoading) return "Loading...";
+
   return (
     <main>
-      <h1 className="text-lg">HI</h1>
+      <CreatePost />
+      {data?.map((post) => (
+        <Posts
+          comments={post.comments}
+          key={post.id}
+          name={post.user.name}
+          avatar={post.user.image}
+          postTitle={post.title}
+          id={post.id}
+        />
+      ))}
     </main>
   );
 }